# apps/admin/src/api/generated/model/createAssistantThreadDtoModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Join The Six API
 * Operations API for Join The Six
 * OpenAPI spec version: 1.0.0
 */

export type CreateAssistantThreadDtoModel =
  (typeof CreateAssistantThreadDtoModel)[keyof typeof CreateAssistantThreadDtoModel];

export const CreateAssistantThreadDtoModel = {
  "openai/gpt-56-luna": "openai/gpt-5.6-luna",
  "openai/gpt-56-terra": "openai/gpt-5.6-terra",
  "google/gemini-36-flash": "google/gemini-3.6-flash",
  "qwen/qwen37-max": "qwen/qwen3.7-max",
} as const;