# apps/backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AssistantThreadDto_Output": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "title": {
            "maxLength": 160,
            "minLength": 1,
            "type": "string"
          },
          "turns": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "assistant": {
                  "additionalProperties": false,
                  "nullable": true,
                  "properties": {
                    "content": {
                      "maxLength": 20000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "role": {
                      "enum": [
                        "assistant"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "role",
                    "content"
                  ],
                  "type": "object"
                },
                "attempt": {
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991,
                  "minimum": 0,
                  "type": "integer"
                },
                "completedAt": {
                  "format": "date-time",
                  "nullable": true,
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "type": "string"
                },
                "createdAt": {
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "type": "string"
                },
                "effort": {
                  "enum": [
                    "low",
                    "medium",
                    "high"
                  ],
                  "type": "string"
                },
                "error": {
                  "additionalProperties": false,
                  "nullable": true,
                  "properties": {
                    "code": {
                      "enum": [
                        "provider_unavailable",
                        "provider_rejected",
                        "generation_failed"
                      ],
                      "type": "string"
                    },
                    "message": {
                      "maxLength": 500,
                      "minLength": 1,
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "type": "object"
                },
                "id": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "model": {
                  "enum": [
                    "openai/gpt-5.6-luna",
                    "openai/gpt-5.6-terra",
                    "google/gemini-3.6-flash",
                    "qwen/qwen3.7-max"
                  ],
                  "type": "string"
                },
                "requestId": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "sequence": {
                  "exclusiveMinimum": true,
                  "maximum": 9007199254740991,
                  "minimum": 0,
                  "type": "integer"
                },
                "startedAt": {
                  "format": "date-time",
                  "nullable": true,
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "type": "string"
                },
                "status": {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed"
                  ],
                  "type": "string"
                },
                "user": {
                  "additionalProperties": false,
                  "properties": {
                    "content": {
                      "maxLength": 20000,
                      "minLength": 1,
                      "type": "string"
                    },
                    "role": {
                      "enum": [
                        "user"
                      ],
                      "type": "string"
                    }
                  },
                  "required": [
                    "role",
                    "content"
                  ],
                  "type": "object"
                }
              },
              "required": [
                "id",
                "requestId",
                "sequence",
                "status",
                "model",
                "effort",
                "user",
                "assistant",
                "error",
                "attempt",
                "createdAt",
                "startedAt",
                "completedAt"
              ],
              "type": "object"
            },
            "minItems": 1,
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "createdAt",
          "updatedAt",
          "turns"
        ],
        "type": "object"
      },
      "AssistantThreadListDto_Output": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "createdAt": {
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "type": "string"
                },
                "id": {
                  "format": "uuid",
                  "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                  "type": "string"
                },
                "lastModel": {
                  "enum": [
                    "openai/gpt-5.6-luna",
                    "openai/gpt-5.6-terra",
                    "google/gemini-3.6-flash",
                    "qwen/qwen3.7-max"
                  ],
                  "type": "string"
                },
                "lastStatus": {
                  "enum": [
                    "queued",
                    "running",
                    "succeeded",
                    "failed"
                  ],
                  "type": "string"
                },
                "title": {
                  "maxLength": 160,
                  "minLength": 1,
                  "type": "string"
                },
                "updatedAt": {
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                  "type": "string"
                }
              },
              "required": [
                "id",
                "title",
                "lastModel",
                "lastStatus",
                "createdAt",
                "updatedAt"
              ],
              "type": "object"
            },
            "maxItems": 50,
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "AssistantTurnDto_Output": {
        "additionalProperties": false,
        "properties": {
          "assistant": {
            "additionalProperties": false,
            "nullable": true,
            "properties": {
              "content": {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              "role": {
                "enum": [
                  "assistant"
                ],
                "type": "string"
              }
            },
            "required": [
              "role",
              "content"
            ],
            "type": "object"
          },
          "attempt": {
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "effort": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "type": "string"
          },
          "error": {
            "additionalProperties": false,
            "nullable": true,
            "properties": {
              "code": {
                "enum": [
                  "provider_unavailable",
                  "provider_rejected",
                  "generation_failed"
                ],
                "type": "string"
              },
              "message": {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "model": {
            "enum": [
              "openai/gpt-5.6-luna",
              "openai/gpt-5.6-terra",
              "google/gemini-3.6-flash",
              "qwen/qwen3.7-max"
            ],
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          },
          "sequence": {
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "minimum": 0,
            "type": "integer"
          },
          "startedAt": {
            "format": "date-time",
            "nullable": true,
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "type": "string"
          },
          "user": {
            "additionalProperties": false,
            "properties": {
              "content": {
                "maxLength": 20000,
                "minLength": 1,
                "type": "string"
              },
              "role": {
                "enum": [
                  "user"
                ],
                "type": "string"
              }
            },
            "required": [
              "role",
              "content"
            ],
            "type": "object"
          }
        },
        "required": [
          "id",
          "requestId",
          "sequence",
          "status",
          "model",
          "effort",
          "user",
          "assistant",
          "error",
          "attempt",
          "createdAt",
          "startedAt",
          "completedAt"
        ],
        "type": "object"
      },
...
      "CreateAssistantThreadDto": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 20000,
            "minLength": 1,
            "type": "string"
          },
          "effort": {
            "default": "low",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "type": "string"
          },
          "model": {
            "enum": [
              "openai/gpt-5.6-luna",
              "openai/gpt-5.6-terra",
              "google/gemini-3.6-flash",
              "qwen/qwen3.7-max"
            ],
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "content"
        ],
        "type": "object"
      },
      "CreateAssistantTurnDto": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 20000,
            "minLength": 1,
            "type": "string"
          },
          "effort": {
            "default": "low",
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "type": "string"
          },
          "model": {
            "enum": [
              "openai/gpt-5.6-luna",
              "openai/gpt-5.6-terra",
              "google/gemini-3.6-flash",
              "qwen/qwen3.7-max"
            ],
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "content"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {},
    "description": "Operations API for Join The Six",
    "title": "Join The Six API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/assistant/threads": {
      "get": {
        "operationId": "listAssistantThreads",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantThreadListDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      },
      "post": {
        "operationId": "createAssistantThread",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssistantThreadDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantThreadDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/threads/{id}": {
      "get": {
        "operationId": "getAssistantThread",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantThreadDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/threads/{id}/turns": {
      "post": {
        "operationId": "createAssistantTurn",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssistantTurnDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/threads/{threadId}/turns/{turnId}": {
      "get": {
        "operationId": "getAssistantTurn",
        "parameters": [
          {
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "turnId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/v1/assistant/threads/{threadId}/turns/{turnId}/retry": {
      "post": {
        "operationId": "retryAssistantTurn",
        "parameters": [
          {
            "in": "path",
            "name": "threadId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "turnId",
            "required": true,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantTurnDto_Output"
                }
              }
            },
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "assistant"
        ]
      }
    },
...
  },
  "servers": [],
...
}